[Decomp][ObjectMgr] Cohesion split: extract instance-template loaders to ObjectMgrInstanceData.cpp
